extracted AppCompatAct and support Frag based classes into module
