+++ AndroidBoundExample/src/main/java/solutions/alterego/androidbound/example/ListViewActivity.java
@@ -5,12 +5,11 @@ import com.codemonkeylabs.fpslibrary.TinyDancer;
 
 import android.os.Bundle;
 
-import solutions.alterego.androidbound.android.support.BindingAppCompatActivity;
 import solutions.alterego.androidbound.example.util.AdvancedAndroidLoggerAdapter;
 import solutions.alterego.androidbound.example.viewmodels.ListViewActivityViewModel;
 import solutions.alterego.androidbound.interfaces.ILogger;
 import solutions.alterego.androidbound.interfaces.IViewBinder;
-
+import solutions.alterego.androidbound.support.android.BindingAppCompatActivity;
 
 public class ListViewActivity extends BindingAppCompatActivity {
