+++ AndroidBoundExample/src/main/java/solutions/alterego/androidbound/example/MainBindingActivity.java
@@ -4,11 +4,11 @@ import com.alterego.advancedandroidlogger.interfaces.IAndroidLogger;
 
 import android.os.Bundle;
 
-import solutions.alterego.androidbound.android.support.BindingAppCompatActivity;
 import solutions.alterego.androidbound.example.util.AdvancedAndroidLoggerAdapter;
 import solutions.alterego.androidbound.example.viewmodels.MainBindingActivityViewModel;
 import solutions.alterego.androidbound.interfaces.ILogger;
 import solutions.alterego.androidbound.interfaces.IViewBinder;
+import solutions.alterego.androidbound.support.android.BindingAppCompatActivity;
 
 public class MainBindingActivity extends BindingAppCompatActivity {
