+++ androidbound-support/src/main/java/solutions/alterego/androidbound/support/android/interfaces/IHasSupportFragmentManager.java
@@ -1,4 +1,4 @@
-package solutions.alterego.androidbound.android.support;
+package solutions.alterego.androidbound.support.android.interfaces;
 
 import android.support.v4.app.FragmentManager;
